Move NumberOfVotersForm to generic components and add story with test
